fix: only use ctx.attr.patches if patches param is not specified (#105)
The currently implementation differs from the docstring here and in upstream skylib:

This rule is intended to be used in the implementation function of
a repository rule. If the parameters "patches", "patch_tool",
"patch_args", "patch_cmds" and "patch_cmds_win" are not specified
then they are taken from "ctx.attr".